feature: update controllers, add annotation fix: update auth

+++ wb-model-entity/src/main/java/com/wangboot/model/entity/controller/RestfulApiController.java
@@ -2,12 +2,12 @@ package com.wangboot.model.entity.controller;
 
 import com.wangboot.core.auth.annotation.RestPermissionAction;
 import com.wangboot.core.auth.authorization.resource.ApiResource;
+import com.wangboot.core.web.exception.NotSupportedComponentException;
 import com.wangboot.model.entity.IRestfulService;
 import com.wangboot.model.entity.IdEntity;
+import com.wangboot.model.entity.RequestConstants;
 import com.wangboot.model.entity.request.IdListBody;
-import com.wangboot.model.entity.utils.EntityUtils;
 import java.io.Serializable;
-import lombok.Generated;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,17 +18,17 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 /**
- * 集成增、删、改、查 Restful 接口的控制器抽象基类
+ * 集成 Restful 接口的控制器抽象基类
  *
  * @param <I> 主键类型
- * @param <S> 服务类型
  * @param <T> 实体类型
+ * @param <S> 服务类型
+ * @author wwtg99
  */
-@Generated
-public abstract class RestfulApiFullController<
+public abstract class RestfulApiController<
         I extends Serializable, T extends IdEntity<I>, S extends IRestfulService<I, T>>
+    extends RestfulBaseController
     implements IRestfulReadController<I, T>, IRestfulWriteController<I, T> {
-
   @Getter @Setter private ApplicationEventPublisher applicationEventPublisher;
 
   @Autowired @Getter private S entityService;
@@ -49,6 +49,9 @@ public abstract class RestfulApiFullController<
   @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_VIEW)
   @NonNull
   public ResponseEntity<?> listApi() {
+    if (!isEnabledApi(ControllerApi.LIST)) {
+      throw new NotSupportedComponentException("Invalid list api for " + this.getClass().getName());
+    }
     return this.listPageResponse();
   }
 
@@ -56,6 +59,10 @@ public abstract class RestfulApiFullController<
   @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_VIEW)
   @NonNull
   public ResponseEntity<?> detailApi(@PathVariable I id) {
+    if (!isEnabledApi(ControllerApi.DETAIL)) {
+      throw new NotSupportedComponentException(
+          "Invalid detail api for " + this.getClass().getName());
+    }
     return this.viewResponse(id);
   }
 
@@ -63,6 +70,10 @@ public abstract class RestfulApiFullController<
   @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_CREATE)
   @NonNull
   public ResponseEntity<?> createApi(@Validated @RequestBody T obj) {
+    if (!isEnabledApi(ControllerApi.CREATE)) {
+      throw new NotSupportedComponentException(
+          "Invalid create api for " + this.getClass().getName());
+    }
     return this.createResponse(obj);
   }
 
@@ -70,7 +81,11 @@ public abstract class RestfulApiFullController<
   @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_UPDATE)
   @NonNull
   public ResponseEntity<?> updateApi(@PathVariable I id, @Validated @RequestBody T obj) {
-    EntityUtils.setEntityIdentifier(obj, id);
+    if (!isEnabledApi(ControllerApi.UPDATE)) {
+      throw new NotSupportedComponentException(
+          "Invalid update api for " + this.getClass().getName());
+    }
+    obj.setId(id);
     return this.updateResponse(obj);
   }
 
@@ -78,6 +93,10 @@ public abstract class RestfulApiFullController<
   @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_DELETE)
   @NonNull
   public ResponseEntity<?> removeApi(@PathVariable I id) {
+    if (!isEnabledApi(ControllerApi.REMOVE)) {
+      throw new NotSupportedComponentException(
+          "Invalid remove api for " + this.getClass().getName());
+    }
     return this.deleteByIdResponse(id);
   }
 
@@ -85,6 +104,32 @@ public abstract class RestfulApiFullController<
   @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_DELETE)
   @NonNull
   public ResponseEntity<?> batchRemoveApi(@Validated @RequestBody IdListBody<I> ids) {
+    if (!isEnabledApi(ControllerApi.BATCH_REMOVE)) {
+      throw new NotSupportedComponentException(
+          "Invalid batch remove api for " + this.getClass().getName());
+    }
     return this.batchDeleteByIdResponse(ids.getIds());
+  }
+
+  @GetMapping("/" + RequestConstants.PATH_ROOT)
+  @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_VIEW)
+  @NonNull
+  public ResponseEntity<?> listRootApi() {
+    if (!isEnabledApi(ControllerApi.LIST_ROOT)) {
+      throw new NotSupportedComponentException(
+          "Invalid list root api for " + this.getClass().getName());
+    }
+    return this.listRootResponse();
+  }
+
+  @GetMapping("/{id}/" + RequestConstants.PATH_CHILDREN)
+  @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_VIEW)
+  @NonNull
+  public ResponseEntity<?> listChildrenApi(@PathVariable I id) {
+    if (!isEnabledApi(ControllerApi.LIST_CHILDREN)) {
+      throw new NotSupportedComponentException(
+          "Invalid list children api for " + this.getClass().getName());
+    }
+    return this.listDirectChildrenResponse(id);
   }
 }
