+++ wb-model-entity/src/main/java/com/wangboot/model/entity/controller/RestfulApiReadWriteController.java
@@ -2,10 +2,11 @@ package com.wangboot.model.entity.controller;
 
 import com.wangboot.core.auth.annotation.RestPermissionAction;
 import com.wangboot.core.auth.authorization.resource.ApiResource;
+import com.wangboot.core.web.exception.NotSupportedComponentException;
 import com.wangboot.model.entity.IRestfulService;
 import com.wangboot.model.entity.IdEntity;
+import com.wangboot.model.entity.RequestConstants;
 import com.wangboot.model.entity.request.IdListBody;
-import com.wangboot.model.entity.utils.EntityUtils;
 import java.io.Serializable;
 import lombok.Getter;
 import lombok.Setter;
@@ -24,6 +25,7 @@ import org.springframework.web.bind.annotation.*;
  * @param <RS> 读服务类型
  * @param <W> 写实体类型
  * @param <WS> 写服务类型
+ * @author wwtg99
  */
 public abstract class RestfulApiReadWriteController<
         I extends Serializable,
@@ -31,7 +33,7 @@ public abstract class RestfulApiReadWriteController<
         RS extends IRestfulService<I, R>,
         W extends IdEntity<I>,
         WS extends IRestfulService<I, W>>
-    implements IRestfulReadWriteController<I, R, W> {
+    extends RestfulBaseController implements IRestfulReadWriteController<I, R, W> {
 
   @Getter @Setter private ApplicationEventPublisher applicationEventPublisher;
 
@@ -55,6 +57,9 @@ public abstract class RestfulApiReadWriteController<
   @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_VIEW)
   @NonNull
   public ResponseEntity<?> listApi() {
+    if (!isEnabledApi(ControllerApi.LIST)) {
+      throw new NotSupportedComponentException("Invalid list api for " + this.getClass().getName());
+    }
     return this.listPageResponse();
   }
 
@@ -62,6 +67,10 @@ public abstract class RestfulApiReadWriteController<
   @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_VIEW)
   @NonNull
   public ResponseEntity<?> detailApi(@PathVariable I id) {
+    if (!isEnabledApi(ControllerApi.DETAIL)) {
+      throw new NotSupportedComponentException(
+          "Invalid detail api for " + this.getClass().getName());
+    }
     return this.viewResponse(id);
   }
 
@@ -69,6 +78,10 @@ public abstract class RestfulApiReadWriteController<
   @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_CREATE)
   @NonNull
   public ResponseEntity<?> createApi(@Validated @RequestBody W obj) {
+    if (!isEnabledApi(ControllerApi.CREATE)) {
+      throw new NotSupportedComponentException(
+          "Invalid create api for " + this.getClass().getName());
+    }
     return this.createResponse(obj);
   }
 
@@ -76,7 +89,11 @@ public abstract class RestfulApiReadWriteController<
   @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_UPDATE)
   @NonNull
   public ResponseEntity<?> updateApi(@PathVariable I id, @Validated @RequestBody W obj) {
-    EntityUtils.setEntityIdentifier(obj, id);
+    if (!isEnabledApi(ControllerApi.UPDATE)) {
+      throw new NotSupportedComponentException(
+          "Invalid update api for " + this.getClass().getName());
+    }
+    obj.setId(id);
     return this.updateResponse(obj);
   }
 
@@ -84,6 +101,10 @@ public abstract class RestfulApiReadWriteController<
   @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_DELETE)
   @NonNull
   public ResponseEntity<?> removeApi(@PathVariable I id) {
+    if (!isEnabledApi(ControllerApi.REMOVE)) {
+      throw new NotSupportedComponentException(
+          "Invalid remove api for " + this.getClass().getName());
+    }
     return this.deleteByIdResponse(id);
   }
 
@@ -91,6 +112,32 @@ public abstract class RestfulApiReadWriteController<
   @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_DELETE)
   @NonNull
   public ResponseEntity<?> batchRemoveApi(@Validated @RequestBody IdListBody<I> ids) {
+    if (!isEnabledApi(ControllerApi.BATCH_REMOVE)) {
+      throw new NotSupportedComponentException(
+          "Invalid batch remove api for " + this.getClass().getName());
+    }
     return this.batchDeleteByIdResponse(ids.getIds());
+  }
+
+  @GetMapping("/" + RequestConstants.PATH_ROOT)
+  @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_VIEW)
+  @NonNull
+  public ResponseEntity<?> listRootApi() {
+    if (!isEnabledApi(ControllerApi.LIST_ROOT)) {
+      throw new NotSupportedComponentException(
+          "Invalid list root api for " + this.getClass().getName());
+    }
+    return this.listRootResponse();
+  }
+
+  @GetMapping("/{id}/" + RequestConstants.PATH_CHILDREN)
+  @RestPermissionAction(ApiResource.REST_PERMISSION_ACTION_VIEW)
+  @NonNull
+  public ResponseEntity<?> listChildrenApi(@PathVariable I id) {
+    if (!isEnabledApi(ControllerApi.LIST_CHILDREN)) {
+      throw new NotSupportedComponentException(
+          "Invalid list children api for " + this.getClass().getName());
+    }
+    return this.listDirectChildrenResponse(id);
   }
 }
